+++ inst/include/dplyr/Result/RowwiseSubset.h
@@ -4,7 +4,7 @@
 #include <tools/ShrinkableVector.h>
 #include <tools/utils.h>
 
-#include <dplyr/check_supported_type.h>
+#include <dplyr/checks.h>
 
 #include <dplyr/Result/GroupedSubsetBase.h>
